Allowing to get annotation-based IC by using key (term)/ value (statistical count or entity list)
I found a mistake in the annotation-based IC function. The numerical values are transferred from children to parent.
That provides an error for the real annotation-based IC. For example, the term A is annotating the
entities 1, 2, and 3. Then, the parent B is annotating the entities 3 and 4. By using counts, A and B have 3 and
2 entities, and using your code, B will have (thanks to the true path rule) a total of 5. That is an error since
A intersect B is equal to 3 (therefore the term B should have a count of 4). My modifications include the following:

    - remove the transfer of counts between children and parents. Then the user need to include a key/value
      where count(parent)>count(children)
    - add a new step to allow users to include a dict() with key representing the keys and the value represent a list of
      entities. For example, with that code we can run the following code:

      from PySLM import *
      ICScores = InformationContent()
      TermStats = {'GO:0140253': ['Prot1', 'Prot2'], 'GO:0009987': ['Prot3'], 'GO:0008150': ['Prot4']}
      ICScores.getIC(['stats'], ['GO:0140253','GO:0009987','GO:0008150'],TermStats=TermStats)

I am not super profident in python, but the code works very well, the modification can be observed during the merge step in git.